Renamed WishListItem Model to Gifts and changed all code tied to the model
